Remove duplicate content from RefactoringCoreMessages.
